change default firepad text
